feat: replace gsap for carousel and ui/ux improv.
